refactor: update Request models to inherit from StandardRequest
